Add display of history graph between branch heads
The option is pretty primitive at the moment. It just displays
git log --graph starting at the newer head and ending when it finds
the SHA1 for the older head. This could obviously stop short if
someone for some reason mentions that SHA in a later commit message.

It also forces color output, as it is assumed that this program
will not be usually piped into a file or another process.